fixed the issue and added a product

+++ nameofapp/db/migrate/20161102144633_add_price_to_products.rb
@@ -1,5 +1,5 @@
 class AddPriceToProducts < ActiveRecord::Migration[5.0]
   def change
-    add_column :products, :price, :decimal
+    add_column :products, :price, :integer
   end
 end
